+++ src/components/create-appointment-modal/SelectLocationsForm.svelte
@@ -60,7 +60,7 @@
 <!-- svelte-ignore a11y-click-events-have-key-events -->
 <form
 	class="grid grid-cols-1 grid-rows-3 gap-6  "
-	id="4-part"
+	id="3-part"
 	on:click|stopPropagation={() => (places = [])}
 	on:submit|preventDefault={handleSubmit}
 >